% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_algo.R
\name{adjust_orient}
\alias{adjust_orient}
\title{Adjusts the orientation of a set of matrices for maximum similarity.}
\usage{
adjust_orient(lambdahat, ref_condit = 1, m)
}
\arguments{
\item{lambdahat}{An array of dimension \code{(p,m,K)}}

\item{ref_condit}{Which of the \code{K} matrices (slices of the array) should be used as reference? (ie: what should everything else be rotated to?)}

\item{m}{how many factors in the dataset?}
}
\value{
An array matching the dimensions of \code{lambdahat}
}
\description{
If matrices have a single column, it checks whether multiplying non-reference conditions by \code{-1} decreases the angle between the reference condition and another condition.
If matrices have more than a single column, it performs a procrustes rotation for maximum similarity.
}
\details{
Uses \code{angle()} which uses \code{acos()}, and therefore may return an error if \code{m=1} and \code{lambdahat[,,k]} and \code{lambdahat[,,j]} are very close to 180 degrees different.
This is relatively unlikely to happen.
}
\examples{
\dontrun{
test_lh <- adjust_orient(covmats_etc$lambdahat, ref_condit = 1, m=2)
}
}
