% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_algo.R
\name{angle}
\alias{angle}
\title{Calculate the angle between two vectors}
\usage{
angle(x1, x2)
}
\arguments{
\item{x1}{A vector or matrix with one column}

\item{x2}{same as x1, same length}
}
\value{
The angle between \code{x1} and \code{x2}.
}
\description{
This function calculates the angle, in degrees, between two vectors.
}
\details{
Uses \code{acos()}, which will sometimes return \code{NaN} if the vectors are very close to 180 degrees apart.
}
\examples{
angle(c(0,1), c(1,0))
angle(c(1.5, 1.5), c(-1,-1))
}
