% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_dta_doc.R
\docType{data}
\name{B_vals}
\alias{B_vals}
\title{B^k values used to generate simulated data}
\format{A list
\describe{
    \item{B_vals[[1]]}{a matrix of dimensions c(30,3), normalized with (ID1).}
     \item{B_vals[[2]]}{a matrix of dimensions c(30,4), normalized with (ID2).}
}}
\usage{
B_vals
}
\description{
a list of length 2, containing matrices of sizes c(30,3) and c(30,4), respectively, containing true values used in data generation
}
\keyword{datasets}
