% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_dta_doc.R
\docType{data}
\name{bluered_palette}
\alias{bluered_palette}
\title{a palette for visualizating reconstructed covariance matrices}
\format{a vector of length 15}
\usage{
bluered_palette
}
\description{
a vector of length 15, with
}
\keyword{datasets}
