% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_fa_viz.R
\name{color_break_vals}
\alias{color_break_vals}
\title{A vector with cutpoints to be used in conjunction with fancyplots}
\usage{
color_break_vals(covmats)
}
\arguments{
\item{covmats}{A list of covariance matrices that will be plotted via fancyplots}
}
\value{
a vector of 16 cutpoints.
}
\description{
Takes as inputs all covariance matrices to be plotted via fancyplots, and returns a sequence of cutpoints so that there are approximately the same number of data points between each cut.
}
\details{
The extreme cutpoints are located at the min and max of all covariance matrices, with intermediate cutpoints spaced approximately at
\code{quantile(, probs = seq(0,1,length=9))} over the negative and positive covariance values separately.
}
\examples{
\dontrun{
 colvals <- color_break_vals(reconstructed$recon_cov)
 }
}
