% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_algo.R
\name{estimate_B}
\alias{estimate_B}
\title{Estimates Bhat via covmat_fa algorithm}
\usage{
estimate_B(Qhat, lambdahat, norm_groups, factsize, design_levels)
}
\arguments{
\item{Qhat}{Estimated Qhat from application of covma_fa algorithm}

\item{lambdahat}{An array of dimensions \code{(p,m,K)}}

\item{norm_groups}{Which of the \code{K} conditions should be normalized together?}

\item{factsize}{Number of latent factors (same as \code{m})}

\item{design_levels}{1 or 2, depending on the experimental design. see vignette for details}
}
\value{
estimated Bhat
}
\description{
Estimates Bhat via covmat_fa algorithm
}
