% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_algo.R
\name{estimate_Q}
\alias{estimate_Q}
\title{Estimates Qhat via covmat_fa algorithm}
\usage{
estimate_Q(lambdahat, design_levels, qadj)
}
\arguments{
\item{lambdahat}{An array of dimensions \code{(p,m,K)}}

\item{design_levels}{parameter indicating normalization structure.}

\item{qadj}{"y" or "n", indicating whether \code{Qhat} should be adjusted w/ the scaling factor {s_ij}}
}
\value{
estimated Qhat
}
\description{
Estimates Qhat via covmat_fa algorithm
}
