% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_dta_doc.R
\docType{data}
\name{Lambda_vals}
\alias{Lambda_vals}
\title{Lambda^k values used to generate simulated data}
\format{A list
\describe{
    \item{Lambda_vals[[1]]}{an array of dimensions c(30,2,3).}
     \item{Lambda_vals[[2]]}{an array of dimensions c(30,2,4).}
}}
\usage{
Lambda_vals
}
\description{
a list of length 2, containing arrays of sizes c(30,2,3) and c(30,2,4), respectively, containing true values used in data generation.
 \code{Lambda[[1]][,,i]} is equivalent to \code{diag(B_vals[[1]][,i])\%\*\%Q_vals}
}
\keyword{datasets}
