% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_fa_viz.R
\name{layout_2x1}
\alias{layout_2x1}
\title{Plots fancyplots for 2x1 experimental design}
\usage{
layout_2x1(base_dim = 5, fancyplots_setup_obj_l2, orient)
}
\arguments{
\item{base_dim}{(optional) an integer value giving the base size for a cell in layout grid.}

\item{fancyplots_setup_obj_l2}{Output from fancyplots_setup. Should be \code{fancyplots_setup_obj$charts} to plot the actual data, or \code{fancyplots_setup_obj$labels} to plot labels}

\item{orient}{\code{orient=1} has both heatmaps in one row; \code{orient=2} has both heatmaps in one column.}
}
\value{
fancyplots for 2x1 experimental design
}
\description{
Plots pseudoLFC bar charts and reconstructed covariance matrices for data having a 2x1 experimental design setup.
}
\examples{
 \dontrun{layout_2x1(fancy[[2]]$charts, orient = 1)}
}
