% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_dta_doc.R
\docType{data}
\name{Q_vals}
\alias{Q_vals}
\title{Q used to generate simulated data}
\format{A matrix of size c(30,2)}
\usage{
Q_vals
}
\description{
true Q values used in data generation
}
\keyword{datasets}
