% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_algo.R
\name{sparse_lambdahat_cardinality}
\alias{sparse_lambdahat_cardinality}
\title{Estimate the number and cardinality of latent factors}
\usage{
sparse_lambdahat_cardinality(step0_obj)
}
\arguments{
\item{step0_obj}{output from \code{step0()}}
}
\value{
A list with the estimated number of factors (factsize_est) and the number of non-zero elements in each factor (col_card)
}
\description{
Estimates the number and cardinality of latent factors by running the leading eigenvector community detection algorithm.
}
\details{
Transforms each covariance matrix into a graph, and uses \code{igraph::leading.eigenvector.community()}, to detect distinct communities.
The smallest number of communities in one covariance matrix is returned as the estimated number of factors.
The sizes of each community are returned as the target column cardinalities.
}
\examples{
#generate the covariance matrices from example datasets
covmats_etc <- vector(mode="list", length=2)
covmats_etc[[1]] <- step0(X[[1]], 2)
covmats_etc[[2]] <- step0(X[[2]], 2)
#calculate the number of factors and number of non-sparse elements for each factor
sparse_params <- vector(mode="list", length=2)
for(i in 1:length(sparse_params)){
 sparse_params[[i]] <- sparse_lambdahat_cardinality(covmats_etc[[i]])
 }
}
