% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_algo.R
\name{step0}
\alias{step0}
\title{Preliminary computations for \code{estimateBkQ}}
\usage{
step0(dta, m = 10)
}
\arguments{
\item{dta}{a list of datasets you wish to use in the covma_fa algorithm}

\item{m}{how many columns do you want in your scaled eigen decomposition?}
}
\value{
A list containing covariance matrices, non-sparse lambdahats (from the scaled eigen-decomposition),
a matrix with percent variance explained by the first \code{m} eigen vectors
and results from the eigen decomposition of the covariance of each dataset.
}
\description{
Returns a list of covariance matrices, non-sparse lambdahats and other useful preliminary items.
}
\examples{
#X[[1]] contains simulated data from a 3x1 experimental design scenario
#X[[2]] contains simulated data from a 2x2 experimental design scenario

covmats_etc <- vector(mode="list", length=2)
covmats_etc[[1]] <- step0(X[[1]], 2)
covmats_etc[[2]] <- step0(X[[2]], 2)
}
