% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_algo.R
\name{update_lambda}
\alias{update_lambda}
\title{Updates lambdahat to remove entries that lead to negative values in Bhat}
\usage{
update_lambda(lambdahat, negB)
}
\arguments{
\item{lambdahat}{An array of dimensions \code{(p,m,K)}}

\item{negB}{parameter indicating normalization structure.}
}
\value{
updated lambdahat, where noisy entries are removed
}
\description{
Updates lambdahat to remove entries that lead to negative values in Bhat
}
