% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_fa_viz.R
\name{vert_bar}
\alias{vert_bar}
\title{Vertical bar chart of pseudo_LFC values}
\usage{
vert_bar(dta, minbar, maxbar)
}
\arguments{
\item{dta}{A vector or matrix with one column, containing data to be plotted}

\item{minbar}{most negative value across all barcharts that will appear in fancyplots}

\item{maxbar}{most positive value across all barcharts that will appear in fancyplots}
}
\value{
a ggplot object suitable for plotting with fancyplots
}
\description{
Produces a vertically oriented bar chart of pseudo_LFC values, suitable for plotting with fancyplots.
}
