% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_fa_viz.R
\name{VP_setup}
\alias{VP_setup}
\title{Helper function that setups the underlying grid layout for use in fancyplots}
\usage{
VP_setup(col_width, row_height)
}
\arguments{
\item{col_width}{A vector of column widths in inches}

\item{row_height}{A vector of row heights in inches}
}
\description{
Helper function that setups the underlying grid layout for use in fancyplots
}
