% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWFM_dta_doc.R
\docType{data}
\name{X}
\alias{X}
\title{simulated data for examples}
\format{A list of length 2, containing lists of length 3 and 4, respectively
\describe{
    \item{X[[1]]}{data from a 3x1 experimental design (\code{X[[1]][[1]]} through \code{X[[1]][[3]]}.}
    \item{X[[2]]}{data from a 2x2 experimental design (\code{X[[2]][[1]]} through \code{X[[2]][[4]]}.}
}}
\usage{
X
}
\description{
A list of lists containing simulated data from a 3x1 experimental design (\code{X[[1]]}), and a 2x2 experimental design (\code{X[[2]]})
}
\keyword{datasets}
